Remove intentional pairs when switching teams
Sucks that I had to move the emisison of events down a level so that now
the teams component is just a message relay between the dev component
and the app component. Sucks how angular2 forces you to re-do events but
in a different way than dom events would do. I guess it's probably a
good thing after all, that it's forcing you to be explicit. But the
event relay thing is still a drag. We could solve all of this by
switching to an "all state at the top of the app" architecture with
nothing but stateless components all the way down below the top-level
app.